feat: set ID and description for VIP-208
